Refactoring: 알람 보내기 부분 로직 수정 (#72)
