+++ backend/src/main/java/com/api/backend/team/data/dto/TeamParticipantsDeleteResponse.java
@@ -5,10 +5,12 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 
 @Getter
 @Setter
 @Builder
+@ToString
 @NoArgsConstructor
 @AllArgsConstructor
 public class TeamParticipantsDeleteResponse {
